Move queue entity/repo to Customers namespace
Prefactoring before starting implementing queue logic
